pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-soho-registry"
version = "0.1.0"
description = "SOHO MCP Server Registry & Auto-Orchestrator"
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "pyyaml>=6.0",
    "jinja2>=3.1",
]

[project.scripts]
mcp-registry = "mcp_registry.cli:cli"
mcp-orchestrate = "mcp_orchestrator.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["mcp_registry", "mcp_orchestrator"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
]
